Mi_proyecto/requirements.txt
django
Pillow
django-crispy-forms
django-filter